Remove redundant cast

+++ Ressy/ResourceName.cs
@@ -41,7 +41,7 @@ public partial class ResourceName
     /// </summary>
     public static ResourceName FromString(string name) => new StringResourceName(name);
 
-    internal static ResourceName FromHandle(nint handle) => (long)handle < 0x10000
+    internal static ResourceName FromHandle(nint handle) => handle < 0x10000
         ? FromCode((int)handle)
         : FromString(NativeHelpers.GetString(handle));
 }
